backend/requirements.txt
Django==3.1.1
django-cors-headers==3.2.0
django-ordered-model==3.3.0
djangorestframework==3.12.4
gunicorn==20.0.4
Markdown==3.1.1
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.1.2
sentry-sdk==0.14.2
python-dotenv==0.12.0
git+https://github.com/mondaylabs/botmother.git@v1.2.2
requests==2.24.0

pep8==1.7.1
autopep8==1.4.4
pycodestyle==2.5.0
